# Conflicts:
#	src/main/java/com/example/webmusic/controller/search/SearchController.java
#	src/main/java/com/example/webmusic/service/search/SearchService.java
#	src/main/java/com/example/webmusic/service/search/SearchServiceImpl.java

+++ src/main/java/com/example/webmusic/service/like/LikeService.java
@@ -6,7 +6,7 @@ import com.example.webmusic.models.like.Like;
 import java.util.List;
 
 public interface LikeService extends IService<Like> {
-    List<Like> getUserLike(int userID);
+    List<Like> getUserLike(long userID);
     int addUserLike(Like like);
     int deleteUserLike(Like like);
 }
